fix: Store 수 조회 API response type 수정

+++ auth-service/src/main/java/com/example/authservice/service/UserService.java
@@ -52,7 +52,7 @@ public class UserService {
     }
 
     // 전체 시스템의 총 유저 수 조회
-    public long getTotalUserCount() {
+    public int getTotalUserCount() {
         return userRepository.getTotalUserCount();
     }
